backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2
python-dotenv
alembic